Move BaseCrud to `core` directory.
